off_t should be signed.

+++ src/arch/i386/include/stdint.h
@@ -1,19 +1,19 @@
 #ifndef STDINT_H
 #define STDINT_H
 
-typedef unsigned           size_t;
-typedef unsigned	   off_t;
+typedef unsigned long		size_t;
+typedef signed long		off_t;
 
-typedef unsigned char      uint8_t;
-typedef unsigned short     uint16_t;
-typedef unsigned long      uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned char		uint8_t;
+typedef unsigned short		uint16_t;
+typedef unsigned long		uint32_t;
+typedef unsigned long long	uint64_t;
 
-typedef signed char        int8_t;
-typedef signed short       int16_t;
-typedef signed long        int32_t;
-typedef signed long long   int64_t;
+typedef signed char		int8_t;
+typedef signed short		int16_t;
+typedef signed long		int32_t;
+typedef signed long long	int64_t;
 
-typedef unsigned long	   physaddr_t;
+typedef unsigned long		physaddr_t;
 
 #endif /* STDINT_H */
